@@ -1,26 +1,32 @@
 buildscript {
     repositories {
         maven { url = 'https://files.minecraftforge.net/maven' }
+        maven { url = 'https://maven.parchmentmc.org' }
         maven { url = 'https://plugins.gradle.org/m2/' }
-        jcenter()
+        maven {
+            url = "https://repo.spongepowered.org/repository/maven-public/"
+            content { includeGroup "org.spongepowered" }
+        }
         mavenCentral()
     }
     dependencies {
         classpath group: 'net.minecraftforge.gradle', name: 'ForgeGradle', version: '5.1.+', changing: true
-        classpath 'gradle.plugin.com.matthewprenger:CurseGradle:1.4.0'
+        classpath group: 'gradle.plugin.com.matthewprenger', name: 'CurseGradle', version: '1.4.0'
+		classpath group: 'org.parchmentmc', name: 'librarian', version: '1.+'
     }
 }
 apply plugin: 'net.minecraftforge.gradle'
 apply plugin: 'com.matthewprenger.cursegradle'
+apply plugin: 'org.parchmentmc.librarian.forgegradle'
 
 version = "${mod_version}"
-group = "${mod_group}"
-archivesBaseName = "${mod_id}-forge"
+group = "top.theillusivec4.comforts"
+archivesBaseName = "comforts-forge"
 
-java.toolchain.languageVersion = JavaLanguageVersion.of(16)
+java.toolchain.languageVersion = JavaLanguageVersion.of(17)
 
 minecraft {
-    mappings channel: 'official', version: '1.17.1'
+    mappings channel: 'parchment', version: '2021.10.31-1.17.1'
 
     runs {
         client {
@@ -73,7 +79,25 @@ repositories {
 }
 
 dependencies {
-    minecraft "net.minecraftforge:forge:${version_forge}"
+    minecraft "net.minecraftforge:forge:${forge_version}"
+}
+
+def replaceResources = tasks.register("replaceResources", Copy) {
+    it.outputs.upToDateWhen { false }
+
+    it.from(sourceSets.main.resources) {
+        include "META-INF/mods.toml"
+        expand 'version': version, 'mc_version_range': mc_version_range, 'forge_version_range': forge_version_range, "loader_version": loader_version
+    }
+    it.into "$buildDir/resources/main/"
+}
+
+processResources {
+    from("CHANGELOG.md")
+    from("../LICENSE")
+    duplicatesStrategy(DuplicatesStrategy.FAIL)
+    exclude('META-INF/mods.toml')
+    configure { finalizedBy(replaceResources) }
 }
 
 sourceSets {
@@ -86,12 +110,12 @@ sourceSets {
 
 jar {
     manifest {
-        attributes(["Specification-Title"     : "${mod_id}",
-                    "Specification-Vendor"    : "${mod_author}",
+        attributes(["Specification-Title"     : "Comforts",
+                    "Specification-Vendor"    : "C4",
                     "Specification-Version"   : "1",
-                    "Implementation-Title"    : "${mod_id}",
-                    "Implementation-Version"  : "${mod_version}",
-                    "Implementation-Vendor"   : "${mod_author}",
+                    "Implementation-Title"    : "Comforts",
+                    "Implementation-Version"  : "${getArchiveVersion()}",
+                    "Implementation-Vendor"   : "C4",
                     "Implementation-Timestamp": new Date().format("yyyy-MM-dd'T'HH:mm:ssZ")],)
     }
 }
@@ -108,10 +132,10 @@ artifacts {
 jar.finalizedBy('reobfJar')
 
 curseforge {
-    apiKey = findProperty('curseKey') ?: 0
 
     project {
-        id = "${curse_id}"
+        apiKey = findProperty('curseKey') ?: 0
+        id = '276951'
         releaseType = "${curse_release}"
         changelogType = 'markdown'
         changelog = file('CHANGELOG.md')
